test: split task executor effect coverage
